leg, hexapod
Updated leg class to add offsets, and set/get offsets and motor IDs. (Considering whether to just do 0-2 for each leg.. or use ids 0-17 or 1-18, and convert it to 0-3.
Added a way to select coxa, femur, tibia and get or set the offsets.
Added some more stuff to hexapod class.
- Added a leg_configs for offsets and servo/leg names/ids.