Latest
Bezier curve kinda checks out, needs more work. I kind of want to use a class instance for de castejau
TODO:
Create a simulation bezier curve is working, but it's not showing how the leg is moving in 3d space. Also unable to replicate the inverse kinematics (still need to implement foward kinematics).
Does not include body kinematics, or it's rotation matrix (Will be useful for when Yun wants to use the accelerometer).