Adjustments
Added code to offset the transition point from the coxa at a different angle than directly from the hexapod center. Allowing the rear/front legs to not be directly in front of the hexapod but rather offset a little. Must ensure no such a high value that the hexapod legs collide.

Also added variable to pod class to adjust hexapod height up and down (Works.)